Merge Day1 and Day2 code into single Day1.go file
Day2 functionality was found to be redundant and was not required as a separate function. Therefore, Day2 code has been incorporated into Day1.go. Related calls in main.go have been updated to reflect this change. Now, 'FindCalibrateValue' accepts a boolean to distinguish between the first and second problem of Day 1. This leads to improved readability and maintainability of the code.